+++ src/dashboard/chat_view.ts
@@ -3,6 +3,11 @@ import * as fs from "fs";
 import * as path from "path";
 import * as _ from "lodash";
 import * as os from "os";
+
+import {
+    Button
+} from "@vscode/webview-ui-toolkit";
+
 import { BoostExtension } from "../extension/BoostExtension";
 import { marked } from "marked";
 import { cleanCellOutput } from "../extension/extensionUtilities";
@@ -15,14 +20,10 @@ import {
 } from "../data/jupyter_notebook";
 import { ControllerOutputType } from "../controllers/controllerOutputTypes";
 import { chatKernelName } from "../controllers/chat_controller";
-import {
-    noProjectOpenMessage,
-    extensionNotFullyActivated,
-    extensionFailedToActivate,
-} from "../data/boostprojectdata_interface";
 import sanitizeHtml from "sanitize-html";
 import { ICellMetadata } from "@jupyterlab/nbformat";
 import { errorToString } from "../utilities/error";
+import { BaseWebviewViewProvider } from "./BaseWebviewViewProvider";
 
 export const aiName = "Sara";
 
@@ -44,23 +45,18 @@ marked.use(markedHighlight({
     }
     }));		
 */
-export class BoostChatViewProvider implements vscode.WebviewViewProvider {
+export class BoostChatViewProvider extends BaseWebviewViewProvider {
     public static readonly viewType = "polyverse-boost-chat-view";
 
-    private _view?: vscode.WebviewView;
     private _chats?: any;
     private _tempFilename?: string;
-    private _context: vscode.ExtensionContext;
     private _activeid = 0;
-    private _boostExtension: BoostExtension;
     private chatService: BoostServiceHelper;
 
-    constructor(
-        private readonly context: vscode.ExtensionContext,
-        private boostExtension: BoostExtension
+    constructor(context: vscode.ExtensionContext,
+            boostExtension: any
     ) {
-        this._context = context;
-        this._boostExtension = boostExtension;
+        super(context, boostExtension, "Chat");
         this.chatService = new BoostServiceHelper(
             "chatService",
             ControllerOutputType.chat,
@@ -68,44 +64,25 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
         );
     }
 
-    public async resolveWebviewView(
-        webviewView: vscode.WebviewView,
-        context: vscode.WebviewViewResolveContext,
-        _token: vscode.CancellationToken
-    ) {
-        try {
-            this._resolveWebviewView(webviewView, context, _token);
-        } catch (e) {
-            boostLogging.error(
-                `Could not refresh ${aiName} Chat View due to ${e}`,
-                false
-            );
-        }
-    }
-
     async _resolveWebviewView(
         webviewView: vscode.WebviewView,
         context: vscode.WebviewViewResolveContext,
         _token: vscode.CancellationToken
     ) {
-        this._view = webviewView;
+        super._resolveWebviewView(webviewView, context, _token);
 
         this._chats = await this._initializeChats();
 
-        webviewView.webview.options = {
-            // Allow scripts in the webview
-            enableScripts: true,
-
-            localResourceRoots: [this.context.extensionUri],
-        };
-
-        webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+        webviewView.webview.html = this._getHtmlForWebview(webviewView.webview, this._boostExtension.getBoostProjectData());
 
         webviewView.webview.onDidReceiveMessage((data) => {
             switch (data.command) {
-                case "newprompt": {
-                    this._activeid = data.chatindex;
-                    this.updatePrompt(data.prompt, data.chatindex, data.showUI);
+                case "new-prompt": {
+                    // if new chatIndex specified, choose it, otherwise use existing active
+                    if (data.chatindex) {
+                        this._activeid = data.chatindex;
+                    }
+                    this._updatePrompt(data.prompt, data.showUI, data.externalResponse);
                     break;
                 }
                 case "add-chat": {
@@ -120,26 +97,14 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
         });
     }
 
-    public refresh() {
-        try {
-            this._refresh();
-        } catch (e) {
-            boostLogging.error(
-                `Could not refresh ${aiName} Chat View due to ${e}`,
-                false
-            );
-        }
-    }
-    _refresh() {
-        if (this._view) {
-            this._view.webview.html = this._getHtmlForWebview(
-                this._view.webview
-            );
-            this._view.show?.(true);
-        }
-    }
+    protected _getHtmlForWebview(webview: vscode.Webview, boostprojectdata: any): string {
 
-    private _getHtmlForWebview(webview: vscode.Webview) {
+        const message = super._getHtmlForWebview(webview, boostprojectdata);
+        if (message) {
+            return message;
+        }
+
+
         const codiconsUri = webview.asWebviewUri(
             vscode.Uri.joinPath(
                 this._context.extensionUri,
@@ -150,13 +115,13 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
             )
         );
         const htmlPathOnDisk = vscode.Uri.joinPath(
-            this.context.extensionUri,
+            this._context.extensionUri,
             "resources",
             "dashboard",
             "chat.html"
         );
         const jsPathOnDisk = vscode.Uri.joinPath(
-            this.context.extensionUri,
+            this._context.extensionUri,
             "out",
             "dashboard",
             "chat",
@@ -170,24 +135,6 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
         const workspaceFolder = vscode.workspace.workspaceFolders
             ? vscode.workspace.workspaceFolders[0]
             : ""; // Get the first workspace folder
-
-        let message;
-
-        if (!this._boostExtension.finishedActivation) {
-            if (!vscode.workspace.workspaceFolders) {
-                message = noProjectOpenMessage;
-            } else {
-                message = extensionNotFullyActivated;
-            }
-        } else if (!this._boostExtension.successfullyActivated) {
-            message = extensionFailedToActivate;
-        } else if (!vscode.workspace.workspaceFolders) {
-            message = noProjectOpenMessage;
-        }
-
-        if (message) {
-            return `<html><body><h3>Boost ${aiName} Chat</h3><p>${message}</p></body></html>`;
-        }
 
         const projectName = workspaceFolder
             ? path.basename(workspaceFolder.uri.fsPath)
@@ -215,16 +162,22 @@ export class BoostChatViewProvider implements vscode.WebviewViewProvider {
         return htmlContent;
     }
 
-    public async updatePrompt(
+    async _updatePrompt(
         prompt: string,
-        index: number,
-        showUI: boolean = true
+        showUI: boolean,
+        externalResponse?: string
     ) {
-        //make a call to the service endpoint with the prompt plus existing context
-        //update the chat view with the response
+        // make a call to the service endpoint with the prompt plus existing context
+        // update the chat view with the response
 
         // we don't save the initial prompts - so we can refresh them each time we send... we only save chat history (persisted)
         try {
+            if (externalResponse) {
+                // save chat response and run cleanup
+                this._addResponse(prompt, externalResponse);
+                return;
+            }
+
             const chatNotebook = new BoostNotebook();
             const tempProcessingCell = new BoostNotebookCell(
                 NotebookCellKind.Markup,